+++ libs/database/src/index.ts
@@ -1,4 +1,4 @@
 export * from './database.module';
 export * from './database.service';
-export * from './entity/user/user.entity';
-export * from './entity/user/user.repository';
+export * from './entites/user/user.entity';
+export * from './entites/user/user.repository';
